Gestion des avis V2

+++ src/main/java/com/api/gesboo/repository/UtilisateurRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UtilisateurRepository extends CrudRepository<Utilisateur,Integer> {
+public interface UtilisateurRepository extends CrudRepository<Utilisateur, Long> {
     Optional<Utilisateur> findByEmail(String email);
 }
